function/requirements.txt
google-cloud-firestore==2.23.0